[routes]create a file for each endpoint
