+++ backend/src/app.ts
@@ -8,7 +8,7 @@ import projectsRouter from "./routes/Projects.routes";
 import projectPostRouter from "./routes/ProjectPosts.routes";
 import projectPostReplyRouter from "./routes/ProjectPostReply.routes";
 import githubRouter from "./routes/Auth.routes";
-import friendshipsRouter from "./routes/Friendship.routes";
+import friendshipsRouter from "./routes/Friendship";
 import userPostsRouter from "./routes/UserPosts.routes";
 import messagesRouter from "./routes/Messages.routes";
 import fileRouter from "./routes/FileUploader.routes";
